Introduce a separate node strictly for s2s tests
Reusing the same node both as a cluster member and a federated XMPP
server across different suites led to a mess in the configuration
options (chaotic entries in `hosts`).

The crux of the problem is that when a new node joins the cluster
(mongoose_cluster:join/1),
the side effect is that its global config entries (defined in ejabberd.cfg,
stored in Mnesia table `global`) overwrite the entries already present
in the cluster.
If the config files between the cluster and the new node match (which
should be the case thanks to the deploy procedure) this is a no-op
(that is nothing changes).
If the config config files differ with regard to global options,
the options from a newly joining member will win (last write wins).

Let's keep things simple - cluster members are just cluster members.
Federated servers are just federated servers.
Don't mix them up.